requirements.txt
# Requirements for running the project

requests
beautifulsoup4
PyYAML
openai
httpx
